backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.20.0
openai>=1.50.0
langgraph>=0.2.0
langchain-core>=0.3.0
gitpython>=3.1.0
aiohttp>=3.10.0
python-multipart>=0.0.9
mcp>=1.0.0
websockets>=13.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
